Move register_handle_wne function to helpers
This wrapper is now used in several standalone workers
so it has been moved to general helpers to prevent code duplication.
